requirements.txt
customtkinter==5.0.5
darkdetect==0.8.0
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
wincertstore==0.2
